requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
python-dotenv
ddgs
requests